# [Java] Check for numeric overflow when parsing numbers from ASCII strings.