--- a/agrona-benchmarks/src/main/java/org/agrona/concurrent/MutableDirectBufferPutIntAsciiBenchmark.java
+++ b/agrona-benchmarks/src/main/java/org/agrona/concurrent/MutableDirectBufferPutIntAsciiBenchmark.java
@@ -35,7 +35,7 @@ public class MutableDirectBufferPutIntAsciiBenchmark
 {
     private static final int CAPACITY = 16;
 
-    @Param({ "-2147483648", "0", "-9182", "27085146", "-123456789", "2147483647" })
+    @Param({ "-2147483648", "-1234567890", "0", "-9182", "27085146", "1999999999", "2147483647" })
     private int value;
 
     private final UnsafeBuffer unsafeBuffer = new UnsafeBuffer(ByteBuffer.allocateDirect(CAPACITY));