--- a/agrona-benchmarks/src/main/java/org/agrona/concurrent/MutableDirectBufferPutLongAsciiBenchmark.java
+++ b/agrona-benchmarks/src/main/java/org/agrona/concurrent/MutableDirectBufferPutLongAsciiBenchmark.java
@@ -35,16 +35,15 @@ public class MutableDirectBufferPutLongAsciiBenchmark
 {
     private static final int CAPACITY = 32;
 
-    @Param(
-        { "-9223372036854775808",
-            "0",
-            "-9182",
-            "123456",
-            "97385146",
-            "10101010101",
-            "-6180362504315475",
-            "9223372036854775807"
-        })
+    @Param({
+        "-9223372036854775808",
+        "-1913372036854775855",
+        "0",
+        "-9182",
+        "27085146",
+        "1010101010101010",
+        "8999999999999999999",
+        "9223372036854775807" })
     private long value;
 
     private final UnsafeBuffer unsafeBuffer = new UnsafeBuffer(ByteBuffer.allocateDirect(CAPACITY));